+++ include/simpledbus/SimpleDBus.h
@@ -3,8 +3,7 @@
 #include "base/Connection.h"
 #include "base/Message.h"
 
+#include "interfaces/Introspectable.h"
+#include "interfaces/ObjectManager.h"
+#include "interfaces/Properties.h"
 #include "interfaces/PropertyHandler.h"
-
-#include "common/ObjectManager.h"
-#include "common/Properties.h"
-#include "common/Introspectable.h"