+++ include/simpledbus/interfaces/ObjectManager.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <simpledbus/base/Interface.h>
+
+#include <functional>
+
+namespace SimpleDBus {
+
+class ObjectManager : public Interface {
+  public:
+    ObjectManager(std::shared_ptr<Connection> conn, std::string bus_name, std::string path);
+    ~ObjectManager();
+
+    // Names are made matching the ones from the DBus specification
+    Holder GetManagedObjects(bool use_callbacks = false);
+    std::function<void(std::string path, Holder options)> InterfacesAdded;
+    std::function<void(std::string path, Holder options)> InterfacesRemoved;
+
+    bool process_received_signal(Message& message);
+};
+
+}  // namespace SimpleDBus