+++ src/advanced/Proxy.cpp
@@ -9,7 +9,7 @@ Proxy::Proxy(std::shared_ptr<Connection> conn, const std::string& bus_name, cons
     : _conn(conn), _bus_name(bus_name), _path(path) {}
 
 std::shared_ptr<Interface> Proxy::interfaces_create(const std::string& name, SimpleDBus::Holder options) {
-    return std::make_unique<Interface>(_conn, _bus_name, _path);
+    return std::make_unique<Interface>(_conn, _bus_name, _path, name);
 }
 
 std::shared_ptr<Proxy> Proxy::path_create(const std::string& path) {