+++ src/base/Interface.cpp
@@ -2,8 +2,9 @@
 
 using namespace SimpleDBus;
 
-Interface::Interface(std::shared_ptr<Connection> conn, const std::string& bus_name, const std::string& path)
-    : _conn(conn), _bus_name(bus_name), _path(path), _loaded(true) {}
+Interface::Interface(std::shared_ptr<Connection> conn, const std::string& bus_name, const std::string& path,
+                     const std::string& interface_name)
+    : _conn(conn), _bus_name(bus_name), _path(path), _interface_name(interface_name), _loaded(true) {}
 
 void Interface::load(Holder options) { _loaded = true; }