+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Testing framework.
 - Exceptions for error detection and handling.
+- Proxy object to generalize path and message handling.
 
 ### Changed
 - Log levels use now a sane naming convention.
 - Default logging level can be set during build time.
 - Holders containing dictionaries can now support any generic key type.
+- All interfaces now inherit from a base class.